+++ spec/rubocop/comment_config_spec.rb
@@ -10,33 +10,34 @@ describe Rubocop::CommentConfig do
       [
         '# encoding: utf-8',
         '',
-        '# rubocop:disable MethodLength',
+        '# rubocop:disable Style/MethodLength',
         'def some_method',
         "  puts 'foo'",
         'end',
-        '# rubocop:enable MethodLength',
+        '# rubocop:enable Style/MethodLength',
         '',
         '# rubocop:disable all',
         'some_method',
         '# rubocop:enable all',
         '',
         "code = 'This is evil.'",
-        'eval(code) # rubocop:disable Eval',
+        'eval(code) # rubocop:disable Lint/Eval',
         "puts 'This is not evil.'",
         '',
         'def some_method',
-        "  puts 'Disabling indented single line' # rubocop:disable LineLength",
+        "  puts 'Disabling indented single line' # rubocop:disable " \
+        'Style/LineLength',
         'end',
         '',
         'string = <<END',
-        'This is a string not a real comment # rubocop:disable Loop',
+        'This is a string not a real comment # rubocop:disable Style/Loop',
         'END',
         '',
-        'foo # rubocop:disable MethodCallParentheses',
+        'foo # rubocop:disable Style/MethodCallParentheses',
         '',
-        '# rubocop:enable Void',
+        '# rubocop:enable Style/Void',
         '',
-        '# rubocop:disable For',
+        '# rubocop:disable Style/For',
         'foo'
       ]
     end
@@ -49,33 +50,34 @@ describe Rubocop::CommentConfig do
     end
 
     it 'supports disabling multiple lines with a pair of directive' do
-      method_length_disabled_lines = disabled_lines_of_cop('MethodLength')
+      method_length_disabled_lines =
+        disabled_lines_of_cop('Style/MethodLength')
       expected_part = (3..6).to_a
       expect(method_length_disabled_lines & expected_part)
         .to eq(expected_part)
     end
 
     it 'supports disabling all lines after a directive' do
-      for_disabled_lines = disabled_lines_of_cop('For')
+      for_disabled_lines = disabled_lines_of_cop('Style/For')
       expected_part = (29..source.size).to_a
       expect(for_disabled_lines & expected_part)
         .to eq(expected_part)
     end
 
     it 'just ignores unpaired enabling directives' do
-      void_disabled_lines = disabled_lines_of_cop('Void')
+      void_disabled_lines = disabled_lines_of_cop('Lint/Void')
       expected_part = (27..source.size).to_a
       expect(void_disabled_lines & expected_part).to be_empty
     end
 
     it 'supports disabling single line with a direcive at end of line' do
-      eval_disabled_lines = disabled_lines_of_cop('Eval')
+      eval_disabled_lines = disabled_lines_of_cop('Lint/Eval')
       expect(eval_disabled_lines).to include(14)
       expect(eval_disabled_lines).not_to include(15)
     end
 
     it 'handles indented single line' do
-      line_length_disabled_lines = disabled_lines_of_cop('LineLength')
+      line_length_disabled_lines = disabled_lines_of_cop('Style/LineLength')
       expect(line_length_disabled_lines).to include(18)
       expect(line_length_disabled_lines).not_to include(19)
     end