+++ spec/rubocop/cop/style/space_inside_block_braces_spec.rb
@@ -8,9 +8,9 @@ describe Rubocop::Cop::Style::SpaceInsideBlockBraces do
   subject(:cop) { described_class.new(config) }
   let(:config) do
     merged = Rubocop::ConfigLoader
-      .default_configuration['SpaceInsideBlockBraces'].merge(cop_config)
-    Rubocop::Config.new('Blocks' => { 'Enabled' => false },
-                        'SpaceInsideBlockBraces' => merged)
+      .default_configuration['Style/SpaceInsideBlockBraces'].merge(cop_config)
+    Rubocop::Config.new('Style/Blocks' => { 'Enabled' => false },
+                        'Style/SpaceInsideBlockBraces' => merged)
   end
   let(:cop_config) do
     {
@@ -147,8 +147,8 @@ describe Rubocop::Cop::Style::SpaceInsideBlockBraces do
 
     context 'and Blocks cop enabled' do
       let(:config) do
-        Rubocop::Config.new('Blocks'                 => { 'Enabled' => true },
-                            'SpaceInsideBlockBraces' => cop_config)
+        Rubocop::Config.new('Style/Blocks' => { 'Enabled' => true },
+                            'Style/SpaceInsideBlockBraces' => cop_config)
       end
 
       it 'does auto-correction for single-line blocks' do